stacks folder restructure and api date version
