+++ server/src/stacks/functionapp/versions.ts
@@ -1,0 +1,3 @@
+export enum FunctionAppAPIVersions {
+  '2020-05-01' = '2020-05-01',
+}
