+++ server/src/stacks/stacks.module.ts
@@ -1,14 +1,12 @@
 import { Module } from '@nestjs/common';
 import { SharedModule } from '../shared/shared.module';
-import { StacksWebAppConfigService } from './stacks.webapp.config.service';
-import { StacksWebAppCreateService } from './stacks.webapp.create.service';
-import { StacksFunctionAppConfigService } from './stacks.functionapp.config.service';
-import { StacksFunctionAppCreateService } from './stacks.functionapp.create.service';
 import { StacksController } from './stacks.controller';
+import { WebAppStacksService20200501 } from './webapp/2020-05-01/stacks.service';
+import { FunctionAppStacksService20200501 } from './functionapp/2020-05-01/stacks.service';
 
 @Module({
   imports: [SharedModule],
   controllers: [StacksController],
-  providers: [StacksFunctionAppConfigService, StacksFunctionAppCreateService, StacksWebAppConfigService, StacksWebAppCreateService],
+  providers: [WebAppStacksService20200501, FunctionAppStacksService20200501],
 })
 export class StacksModule {}
