--- server/src/stacks/stacks.webapp.config.service.ts
@@ -1,2111 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { WebAppConfigStack } from './stacks';
-
-@Injectable()
-export class StacksWebAppConfigService {
-  private _aspNetWindowsStacks: WebAppConfigStack = {
-    id: null,
-    name: 'aspnet',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
-    properties: {
-      name: 'aspnet',
-      display: 'ASP.NET',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: 'V4.7',
-          runtimeVersion: 'v4.0',
-          isDefault: true,
-          minorVersions: [],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'V3.5',
-          runtimeVersion: 'v2.0',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _nodeWindowsStacks: WebAppConfigStack = {
-    id: null,
-    name: 'node',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
-    properties: {
-      name: 'node',
-      display: 'Node',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '0.6',
-          runtimeVersion: '0.6',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '0.6.20',
-              runtimeVersion: '0.6.20',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '0.8',
-          runtimeVersion: '0.8',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '0.8.2',
-              runtimeVersion: '0.8.2',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.8.19',
-              runtimeVersion: '0.8.19',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.8.26',
-              runtimeVersion: '0.8.26',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.8.27',
-              runtimeVersion: '0.8.27',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.8.28',
-              runtimeVersion: '0.8.28',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '0.10',
-          runtimeVersion: '0.10',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '0.10.5',
-              runtimeVersion: '0.10.5',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.18',
-              runtimeVersion: '0.10.18',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.21',
-              runtimeVersion: '0.10.21',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.24',
-              runtimeVersion: '0.10.24',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.28',
-              runtimeVersion: '0.10.28',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.29',
-              runtimeVersion: '0.10.29',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.31',
-              runtimeVersion: '0.10.31',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.32',
-              runtimeVersion: '0.10.32',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.40',
-              runtimeVersion: '0.10.40',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.10.5',
-              runtimeVersion: '0.10.5',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '0.12',
-          runtimeVersion: '0.12',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '0.12.0',
-              runtimeVersion: '0.12.0',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.12.2',
-              runtimeVersion: '0.12.2',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.12.3',
-              runtimeVersion: '0.12.3',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '0.12.6',
-              runtimeVersion: '0.12.6',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '4.8',
-          runtimeVersion: '4.8',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.5',
-          runtimeVersion: '6.5',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6.5.0',
-              runtimeVersion: '6.5.0',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.9',
-          runtimeVersion: '6.9',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6.9.1',
-              runtimeVersion: '6.9.1',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '6.9.5',
-              runtimeVersion: '6.9.5',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.12',
-          runtimeVersion: '6.12',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.10',
-          runtimeVersion: '7.10',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '7.10.1',
-              runtimeVersion: '7.10.1',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.1',
-          runtimeVersion: '8.1',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.1.4',
-              runtimeVersion: '8.1.4',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.4',
-          runtimeVersion: '8.4',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.5',
-          runtimeVersion: '8.5',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.9',
-          runtimeVersion: '8.9',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.10',
-          runtimeVersion: '8.10',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.11',
-          runtimeVersion: '8.11',
-          isDefault: true,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.0',
-          runtimeVersion: '10.0',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.0.0',
-              runtimeVersion: '10.0.0',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.6',
-          runtimeVersion: '10.6',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.6.0',
-              runtimeVersion: '10.6.0',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.14',
-          runtimeVersion: '10.14',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.14.0',
-              runtimeVersion: '10.14.0',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.15',
-          runtimeVersion: '10.15',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.15.2',
-              runtimeVersion: '10.15.2',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _pythonWindowsStacks: WebAppConfigStack = {
-    id: null,
-    name: 'python',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
-    properties: {
-      name: 'python',
-      display: 'Python',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '2.7',
-          runtimeVersion: '2.7',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '2.7',
-              runtimeVersion: '2.7.3',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '3.6',
-          runtimeVersion: '3.4',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '3.6',
-              runtimeVersion: '3.4.0',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _phpWindowsStacks: WebAppConfigStack = {
-    id: null,
-    name: 'php',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
-    properties: {
-      name: 'php',
-      display: 'PHP',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '5.6',
-          runtimeVersion: '5.6',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.0',
-          runtimeVersion: '7.0',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.1',
-          runtimeVersion: '7.1',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.2',
-          runtimeVersion: '7.2',
-          isDefault: true,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.3',
-          runtimeVersion: '7.3',
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _netCoreWindowsStacks: WebAppConfigStack = {
-    id: null,
-    name: 'dotnetcore',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
-    properties: {
-      name: 'dotnetcore',
-      display: '.NET Core',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '1.0',
-          runtimeVersion: null,
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '1.1',
-          runtimeVersion: null,
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.1',
-          runtimeVersion: null,
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.2',
-          runtimeVersion: null,
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '3.0',
-          runtimeVersion: null,
-          isDefault: false,
-          minorVersions: [],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _javaWindowsStacks: WebAppConfigStack = {
-    id: null,
-    name: 'java',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
-    properties: {
-      name: 'java',
-      display: 'Java',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '1.7',
-          runtimeVersion: '1.7',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '1.7.0_51 (Oracle)',
-              runtimeVersion: '1.7.0_51',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.7.0_71 (Oracle)',
-              runtimeVersion: '1.7.0_71',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.7.0_80 (Oracle)',
-              runtimeVersion: '1.7.0_80',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.7.0_191 (Azul)',
-              runtimeVersion: '1.7.0_191_ZULU',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.7.0_222 (Azul)',
-              runtimeVersion: '1.7.0_222_ZULU',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: true,
-          isHidden: false,
-        },
-        {
-          displayVersion: '1.8',
-          runtimeVersion: '1.8',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '1.8.0_25 (Oracle)',
-              runtimeVersion: '1.8.0_25',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_60 (Oracle)',
-              runtimeVersion: '1.8.0_60',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_73 (Oracle)',
-              runtimeVersion: '1.8.0_73',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_92 (Azul)',
-              runtimeVersion: '1.8.0_92',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_102 (Azul)',
-              runtimeVersion: '1.8.0_102',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_111 (Oracle)',
-              runtimeVersion: '1.8.0_111',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_144 (Azul)',
-              runtimeVersion: '1.8.0_144',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_172 (Oracle)',
-              runtimeVersion: '1.8.0_172',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_172 (Azul)',
-              runtimeVersion: '1.8.0_172_ZULU',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_181 (Oracle)',
-              runtimeVersion: '1.8.0_181',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_181 (Azul)',
-              runtimeVersion: '1.8.0_181_ZULU',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_202 (Oracle)',
-              runtimeVersion: '1.8.0_202',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_202 (Azul)',
-              runtimeVersion: '1.8.0_202_ZULU',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '1.8.0_212 (Azul)',
-              runtimeVersion: '1.8.0_212_ZULU',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '11',
-          runtimeVersion: '11',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '11.0.2 (Azul)',
-              runtimeVersion: '11.0.2_ZULU',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '11.0.3 (Azul)',
-              runtimeVersion: '11.0.3_ZULU',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _javaContainerWindowsStacks: WebAppConfigStack = {
-    id: null,
-    name: 'javaContainers',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
-    properties: {
-      name: 'javaContainers',
-      display: 'Java',
-      dependency: 'java',
-      majorVersions: [],
-      frameworks: [
-        {
-          name: 'tomcat',
-          display: 'Tomcat',
-          dependency: null,
-          majorVersions: [
-            {
-              displayVersion: '7.0',
-              runtimeVersion: '7.0',
-              isDefault: false,
-              minorVersions: [
-                {
-                  displayVersion: '7.0.50',
-                  runtimeVersion: '7.0.50',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '7.0.62',
-                  runtimeVersion: '7.0.62',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '7.0.81',
-                  runtimeVersion: '7.0.81',
-                  isDefault: true,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '7.0.94',
-                  runtimeVersion: '7.0.94',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-              ],
-              applicationInsights: false,
-              isPreview: false,
-              isDeprecated: true,
-              isHidden: false,
-            },
-            {
-              displayVersion: '8.0',
-              runtimeVersion: '8.0',
-              isDefault: false,
-              minorVersions: [
-                {
-                  displayVersion: '8.0.23',
-                  runtimeVersion: '8.0.23',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '8.0.46',
-                  runtimeVersion: '8.0.46',
-                  isDefault: true,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '8.0.53',
-                  runtimeVersion: '8.0.53',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-              ],
-              applicationInsights: false,
-              isPreview: false,
-              isDeprecated: true,
-              isHidden: false,
-            },
-            {
-              displayVersion: '8.5',
-              runtimeVersion: '8.5',
-              isDefault: false,
-              minorVersions: [
-                {
-                  displayVersion: '8.5.6',
-                  runtimeVersion: '8.5.6',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '8.5.20',
-                  runtimeVersion: '8.5.20',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '8.5.31',
-                  runtimeVersion: '8.5.31',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '8.5.34',
-                  runtimeVersion: '8.5.34',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '8.5.37',
-                  runtimeVersion: '8.5.37',
-                  isDefault: true,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '8.5.42',
-                  runtimeVersion: '8.5.42',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-              ],
-              applicationInsights: false,
-              isPreview: false,
-              isDeprecated: false,
-              isHidden: false,
-            },
-            {
-              displayVersion: '9.0',
-              runtimeVersion: '9.0',
-              isDefault: true,
-              minorVersions: [
-                {
-                  displayVersion: '9.0.0',
-                  runtimeVersion: '9.0.0',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '9.0.8',
-                  runtimeVersion: '9.0.8',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '9.0.12',
-                  runtimeVersion: '9.0.12',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '9.0.14',
-                  runtimeVersion: '9.0.14',
-                  isDefault: true,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '9.0.21',
-                  runtimeVersion: '9.0.21',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-              ],
-              applicationInsights: false,
-              isPreview: false,
-              isDeprecated: false,
-              isHidden: false,
-            },
-          ],
-        },
-        {
-          name: 'jetty',
-          display: 'Jetty',
-          dependency: null,
-          majorVersions: [
-            {
-              displayVersion: '9.1',
-              runtimeVersion: '9.1',
-              isDefault: false,
-              minorVersions: [
-                {
-                  displayVersion: '9.1.0',
-                  runtimeVersion: '9.1.0.20131115',
-                  isDefault: true,
-                  isRemoteDebuggingEnabled: false,
-                },
-              ],
-              applicationInsights: false,
-              isPreview: false,
-              isDeprecated: false,
-              isHidden: false,
-            },
-            {
-              displayVersion: '9.3',
-              runtimeVersion: '9.3',
-              isDefault: true,
-              minorVersions: [
-                {
-                  displayVersion: '9.3.13',
-                  runtimeVersion: '9.3.13.20161014',
-                  isDefault: false,
-                  isRemoteDebuggingEnabled: false,
-                },
-                {
-                  displayVersion: '9.3.25',
-                  runtimeVersion: '9.3.25.20180904',
-                  isDefault: true,
-                  isRemoteDebuggingEnabled: false,
-                },
-              ],
-              applicationInsights: false,
-              isPreview: false,
-              isDeprecated: false,
-              isHidden: false,
-            },
-          ],
-        },
-        {
-          name: 'java',
-          display: 'Java',
-          dependency: null,
-          majorVersions: [
-            {
-              displayVersion: 'SE',
-              runtimeVersion: 'SE',
-              isDefault: false,
-              minorVersions: [],
-              applicationInsights: false,
-              isPreview: false,
-              isDeprecated: false,
-              isHidden: false,
-            },
-          ],
-        },
-      ],
-      isDeprecated: null,
-    },
-  };
-
-  private _rubyLinuxStacks: WebAppConfigStack = {
-    id: null,
-    name: 'ruby',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Linux',
-    properties: {
-      name: 'ruby',
-      display: 'Ruby',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '2.3',
-          runtimeVersion: 'RUBY|2.3',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '2.3.3',
-              runtimeVersion: 'RUBY|2.3.3',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '2.3.8',
-              runtimeVersion: 'RUBY|2.3.8',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.4',
-          runtimeVersion: 'RUBY|2.4',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '2.4.5',
-              runtimeVersion: 'RUBY|2.4.5',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.5',
-          runtimeVersion: 'RUBY|2.5',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '2.5.5',
-              runtimeVersion: 'RUBY|2.5.5',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.6',
-          runtimeVersion: 'RUBY|2.6',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '2.6.2',
-              runtimeVersion: 'RUBY|2.6.2',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _nodeLinuxStacks: WebAppConfigStack = {
-    id: null,
-    name: 'node',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Linux',
-    properties: {
-      name: 'node',
-      display: 'Node',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: 'LTS',
-          runtimeVersion: 'NODE|lts',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: 'LTS',
-              runtimeVersion: 'NODE|lts',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '12 LTS',
-          runtimeVersion: 'NODE|12-lts',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '12 LTS',
-              runtimeVersion: 'NODE|12-lts',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10 LTS',
-          runtimeVersion: 'NODE|10-lts',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10 LTS',
-              runtimeVersion: 'NODE|10-lts',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8 LTS',
-          runtimeVersion: 'NODE|8-lts',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8',
-              runtimeVersion: 'NODE|8-lts',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6 LTS',
-          runtimeVersion: 'NODE|6-lts',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6',
-              runtimeVersion: 'NODE|6-lts',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '4.4',
-          runtimeVersion: 'NODE|4.4',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '4.4',
-              runtimeVersion: 'NODE|4.4',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '4.5',
-          runtimeVersion: 'NODE|4.5',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '4.5',
-              runtimeVersion: 'NODE|4.5',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '4.8',
-          runtimeVersion: 'NODE|4.8',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '4.8',
-              runtimeVersion: 'NODE|4.8.7',
-              isDefault: false,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.2',
-          runtimeVersion: 'NODE|6.2',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6.2',
-              runtimeVersion: 'NODE|6.2',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.6',
-          runtimeVersion: 'NODE|6.6',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6.6',
-              runtimeVersion: 'NODE|6.6',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.9',
-          runtimeVersion: 'NODE|6.9',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6.9',
-              runtimeVersion: 'NODE|6.9',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.10',
-          runtimeVersion: 'NODE|6.10',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6.10',
-              runtimeVersion: 'NODE|6.10',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '6.11',
-          runtimeVersion: 'NODE|6.11',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '6.11',
-              runtimeVersion: 'NODE|6.11',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.0',
-          runtimeVersion: 'NODE|8.0',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.0',
-              runtimeVersion: 'NODE|8.0',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.1',
-          runtimeVersion: 'NODE|8.1',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.1',
-              runtimeVersion: 'NODE|8.1',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.2',
-          runtimeVersion: 'NODE|8.2',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.2',
-              runtimeVersion: 'NODE|8.2',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.8',
-          runtimeVersion: 'NODE|8.8',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.8',
-              runtimeVersion: 'NODE|8.8',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.9',
-          runtimeVersion: 'NODE|8.9',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.9',
-              runtimeVersion: 'NODE|8.9',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.11',
-          runtimeVersion: 'NODE|8.11',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.11',
-              runtimeVersion: 'NODE|8.11',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '8.12',
-          runtimeVersion: 'NODE|8.12',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.12',
-              runtimeVersion: 'NODE|8.12',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '9.4',
-          runtimeVersion: 'NODE|9.4',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '9.4',
-              runtimeVersion: 'NODE|9.4',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.1',
-          runtimeVersion: 'NODE|10.1',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.1',
-              runtimeVersion: 'NODE|10.1',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.10',
-          runtimeVersion: 'NODE|10.10',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.10',
-              runtimeVersion: 'NODE|10.10',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.12',
-          runtimeVersion: 'NODE|10.12',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.12',
-              runtimeVersion: 'NODE|10.12',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.14',
-          runtimeVersion: 'NODE|10.14',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.14',
-              runtimeVersion: 'NODE|10.14',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '10.16',
-          runtimeVersion: 'NODE|10.16',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '10.16',
-              runtimeVersion: 'NODE|10.16',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '12.9',
-          runtimeVersion: 'NODE|12.9',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '12.9',
-              runtimeVersion: 'NODE|12.9',
-              isDefault: true,
-              isRemoteDebuggingEnabled: true,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _phpLinuxStacks: WebAppConfigStack = {
-    id: null,
-    name: 'php',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Linux',
-    properties: {
-      name: 'php',
-      display: 'PHP',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '5.6',
-          runtimeVersion: 'PHP|5.6',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '5.6-apache',
-              runtimeVersion: 'PHP|5.6',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '5.6-apache-xdebug',
-              runtimeVersion: 'PHP|5.6',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.0',
-          runtimeVersion: 'PHP|7.0',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '7.0-apache',
-              runtimeVersion: 'PHP|7.0',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '7.0-apache-xdebug',
-              runtimeVersion: 'PHP|7.0',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.2',
-          runtimeVersion: 'PHP|7.2',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '7.2-apache',
-              runtimeVersion: 'PHP|7.2',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '7.2-apache-xdebug',
-              runtimeVersion: 'PHP|7.2',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '7.3',
-          runtimeVersion: 'PHP|7.3',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '7.3-apache',
-              runtimeVersion: 'PHP|7.3',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '7.3-apache-xdebug',
-              runtimeVersion: 'PHP|7.3',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-            {
-              displayVersion: '7.3-apache',
-              runtimeVersion: 'PHP|7.3',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _netCoreLinuxStacks: WebAppConfigStack = {
-    id: null,
-    name: 'dotnetcore',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Linux',
-    properties: {
-      name: 'dotnetcore',
-      display: '.NET Core',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '1.0',
-          runtimeVersion: 'DOTNETCORE|1.0',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '1.0.14',
-              runtimeVersion: 'DOTNETCORE|1.0',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '1.1',
-          runtimeVersion: 'DOTNETCORE|1.1',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '1.1.11',
-              runtimeVersion: 'DOTNETCORE|1.1',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.0',
-          runtimeVersion: 'DOTNETCORE|2.0',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '2.0.9',
-              runtimeVersion: 'DOTNETCORE|2.0',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.1',
-          runtimeVersion: 'DOTNETCORE|2.1',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '2.1.8',
-              runtimeVersion: 'DOTNETCORE|2.1',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.2',
-          runtimeVersion: 'DOTNETCORE|2.2',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '2.2.2',
-              runtimeVersion: 'DOTNETCORE|2.2',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '3.0',
-          runtimeVersion: 'DOTNETCORE|3.0',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '3.0.0',
-              runtimeVersion: 'DOTNETCORE|3.0',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'LTS',
-          runtimeVersion: 'DOTNETCORE|LTS',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: 'LTS',
-              runtimeVersion: 'DOTNETCORE|LTS',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'Latest',
-          runtimeVersion: 'DOTNETCORE|Latest',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: 'Latest',
-              runtimeVersion: 'DOTNETCORE|LATEST',
-              isDefault: false,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _java8LinuxStacks: WebAppConfigStack = {
-    id: null,
-    name: 'java8',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Linux',
-    properties: {
-      name: 'java8',
-      display: 'Java 8',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: 'Tomcat 8.5',
-          runtimeVersion: 'TOMCAT|8.5-jre8',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.5',
-              runtimeVersion: 'TOMCAT|8.5-jre8',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'Tomcat 9.0',
-          runtimeVersion: 'TOMCAT|9.0-jre8',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '9.0',
-              runtimeVersion: 'TOMCAT|9.0-jre8',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'Java SE',
-          runtimeVersion: 'JAVA|8-jre8',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '8',
-              runtimeVersion: 'JAVA|8-jre8',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'WildFly 14 - Preview',
-          runtimeVersion: 'WILDFLY|14-jre8',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '14',
-              runtimeVersion: 'WILDFLY|14-jre8',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _java11LinuxStacks: WebAppConfigStack = {
-    id: null,
-    name: 'java11',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Linux',
-    properties: {
-      name: 'java11',
-      display: 'Java 11',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: 'Tomcat 8.5',
-          runtimeVersion: 'TOMCAT|8.5-java11',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '8.5',
-              runtimeVersion: 'TOMCAT|8.5-java11',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'Tomcat 9.0',
-          runtimeVersion: 'TOMCAT|9.0-java11',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '9.0',
-              runtimeVersion: 'TOMCAT|9.0-java11',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: 'Java SE',
-          runtimeVersion: 'JAVA|11-java11',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '11',
-              runtimeVersion: 'JAVA|11-java11',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: true,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  private _pythonLinuxStacks: WebAppConfigStack = {
-    id: null,
-    name: 'python',
-    type: 'Microsoft.Web/availableStacks?osTypeSelected=Linux',
-    properties: {
-      name: 'python',
-      display: 'Python',
-      dependency: null,
-      majorVersions: [
-        {
-          displayVersion: '3.8',
-          runtimeVersion: 'PYTHON|3.8',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '3.8',
-              runtimeVersion: 'PYTHON|3.8',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '3.7',
-          runtimeVersion: 'PYTHON|3.7',
-          isDefault: true,
-          minorVersions: [
-            {
-              displayVersion: '3.7',
-              runtimeVersion: 'PYTHON|3.7',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '3.6',
-          runtimeVersion: 'PYTHON|3.6',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '3.6',
-              runtimeVersion: 'PYTHON|3.6',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-        {
-          displayVersion: '2.7',
-          runtimeVersion: 'PYTHON|2.7',
-          isDefault: false,
-          minorVersions: [
-            {
-              displayVersion: '2.7',
-              runtimeVersion: 'PYTHON|2.7',
-              isDefault: true,
-              isRemoteDebuggingEnabled: false,
-            },
-          ],
-          applicationInsights: false,
-          isPreview: false,
-          isDeprecated: false,
-          isHidden: false,
-        },
-      ],
-      frameworks: [],
-      isDeprecated: null,
-    },
-  };
-
-  getStacks(os?: 'linux' | 'windows'): WebAppConfigStack[] {
-    const windowsStacks = [
-      this._aspNetWindowsStacks,
-      this._nodeWindowsStacks,
-      this._pythonWindowsStacks,
-      this._phpWindowsStacks,
-      this._netCoreWindowsStacks,
-      this._javaWindowsStacks,
-      this._javaContainerWindowsStacks,
-    ];
-
-    const linuxStacks = [
-      this._nodeLinuxStacks,
-      this._pythonLinuxStacks,
-      this._phpLinuxStacks,
-      this._netCoreLinuxStacks,
-      this._rubyLinuxStacks,
-      this._java8LinuxStacks,
-      this._java11LinuxStacks,
-    ];
-
-    if (os === 'linux') {
-      return linuxStacks;
-    } else if (os === 'windows') {
-      return windowsStacks;
-    } else {
-      return windowsStacks.concat(linuxStacks);
-    }
-  }
-}
