+++ server/src/stacks/webapp/2020-05-01/stacks/config/windows/dotnetCore.ts
@@ -1,0 +1,66 @@
+import { WebAppConfigStack } from '../../../stack.model';
+
+export const dotnetCoreWindowsConfigStack: WebAppConfigStack = {
+  id: null,
+  name: 'dotnetcore',
+  type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
+  properties: {
+    name: 'dotnetcore',
+    display: '.NET Core',
+    dependency: null,
+    majorVersions: [
+      {
+        displayVersion: '1.0',
+        runtimeVersion: null,
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '1.1',
+        runtimeVersion: null,
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '2.1',
+        runtimeVersion: null,
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '2.2',
+        runtimeVersion: null,
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '3.0',
+        runtimeVersion: null,
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+    ],
+    frameworks: [],
+    isDeprecated: null,
+  },
+};
