+++ server/src/stacks/webapp/2020-05-01/stacks/config/windows/java.ts
@@ -1,0 +1,175 @@
+import { WebAppConfigStack } from '../../../stack.model';
+
+export const javaWindowsConfigStack: WebAppConfigStack = {
+  id: null,
+  name: 'java',
+  type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
+  properties: {
+    name: 'java',
+    display: 'Java',
+    dependency: null,
+    majorVersions: [
+      {
+        displayVersion: '1.7',
+        runtimeVersion: '1.7',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '1.7.0_51 (Oracle)',
+            runtimeVersion: '1.7.0_51',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.7.0_71 (Oracle)',
+            runtimeVersion: '1.7.0_71',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.7.0_80 (Oracle)',
+            runtimeVersion: '1.7.0_80',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.7.0_191 (Azul)',
+            runtimeVersion: '1.7.0_191_ZULU',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.7.0_222 (Azul)',
+            runtimeVersion: '1.7.0_222_ZULU',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: true,
+        isHidden: false,
+      },
+      {
+        displayVersion: '1.8',
+        runtimeVersion: '1.8',
+        isDefault: true,
+        minorVersions: [
+          {
+            displayVersion: '1.8.0_25 (Oracle)',
+            runtimeVersion: '1.8.0_25',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_60 (Oracle)',
+            runtimeVersion: '1.8.0_60',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_73 (Oracle)',
+            runtimeVersion: '1.8.0_73',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_92 (Azul)',
+            runtimeVersion: '1.8.0_92',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_102 (Azul)',
+            runtimeVersion: '1.8.0_102',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_111 (Oracle)',
+            runtimeVersion: '1.8.0_111',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_144 (Azul)',
+            runtimeVersion: '1.8.0_144',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_172 (Oracle)',
+            runtimeVersion: '1.8.0_172',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_172 (Azul)',
+            runtimeVersion: '1.8.0_172_ZULU',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_181 (Oracle)',
+            runtimeVersion: '1.8.0_181',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_181 (Azul)',
+            runtimeVersion: '1.8.0_181_ZULU',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_202 (Oracle)',
+            runtimeVersion: '1.8.0_202',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_202 (Azul)',
+            runtimeVersion: '1.8.0_202_ZULU',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '1.8.0_212 (Azul)',
+            runtimeVersion: '1.8.0_212_ZULU',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '11',
+        runtimeVersion: '11',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '11.0.2 (Azul)',
+            runtimeVersion: '11.0.2_ZULU',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '11.0.3 (Azul)',
+            runtimeVersion: '11.0.3_ZULU',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+    ],
+    frameworks: [],
+    isDeprecated: null,
+  },
+};
