+++ server/src/stacks/webapp/2020-05-01/stacks/config/windows/javaContainer.ts
@@ -1,0 +1,239 @@
+import { WebAppConfigStack } from '../../../stack.model';
+
+export const javaContainerWindowsConfigStack: WebAppConfigStack = {
+  id: null,
+  name: 'javaContainers',
+  type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
+  properties: {
+    name: 'javaContainers',
+    display: 'Java',
+    dependency: 'java',
+    majorVersions: [],
+    frameworks: [
+      {
+        name: 'tomcat',
+        display: 'Tomcat',
+        dependency: null,
+        majorVersions: [
+          {
+            displayVersion: '7.0',
+            runtimeVersion: '7.0',
+            isDefault: false,
+            minorVersions: [
+              {
+                displayVersion: '7.0.50',
+                runtimeVersion: '7.0.50',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '7.0.62',
+                runtimeVersion: '7.0.62',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '7.0.81',
+                runtimeVersion: '7.0.81',
+                isDefault: true,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '7.0.94',
+                runtimeVersion: '7.0.94',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+            ],
+            applicationInsights: false,
+            isPreview: false,
+            isDeprecated: true,
+            isHidden: false,
+          },
+          {
+            displayVersion: '8.0',
+            runtimeVersion: '8.0',
+            isDefault: false,
+            minorVersions: [
+              {
+                displayVersion: '8.0.23',
+                runtimeVersion: '8.0.23',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '8.0.46',
+                runtimeVersion: '8.0.46',
+                isDefault: true,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '8.0.53',
+                runtimeVersion: '8.0.53',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+            ],
+            applicationInsights: false,
+            isPreview: false,
+            isDeprecated: true,
+            isHidden: false,
+          },
+          {
+            displayVersion: '8.5',
+            runtimeVersion: '8.5',
+            isDefault: false,
+            minorVersions: [
+              {
+                displayVersion: '8.5.6',
+                runtimeVersion: '8.5.6',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '8.5.20',
+                runtimeVersion: '8.5.20',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '8.5.31',
+                runtimeVersion: '8.5.31',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '8.5.34',
+                runtimeVersion: '8.5.34',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '8.5.37',
+                runtimeVersion: '8.5.37',
+                isDefault: true,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '8.5.42',
+                runtimeVersion: '8.5.42',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+            ],
+            applicationInsights: false,
+            isPreview: false,
+            isDeprecated: false,
+            isHidden: false,
+          },
+          {
+            displayVersion: '9.0',
+            runtimeVersion: '9.0',
+            isDefault: true,
+            minorVersions: [
+              {
+                displayVersion: '9.0.0',
+                runtimeVersion: '9.0.0',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '9.0.8',
+                runtimeVersion: '9.0.8',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '9.0.12',
+                runtimeVersion: '9.0.12',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '9.0.14',
+                runtimeVersion: '9.0.14',
+                isDefault: true,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '9.0.21',
+                runtimeVersion: '9.0.21',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+            ],
+            applicationInsights: false,
+            isPreview: false,
+            isDeprecated: false,
+            isHidden: false,
+          },
+        ],
+      },
+      {
+        name: 'jetty',
+        display: 'Jetty',
+        dependency: null,
+        majorVersions: [
+          {
+            displayVersion: '9.1',
+            runtimeVersion: '9.1',
+            isDefault: false,
+            minorVersions: [
+              {
+                displayVersion: '9.1.0',
+                runtimeVersion: '9.1.0.20131115',
+                isDefault: true,
+                isRemoteDebuggingEnabled: false,
+              },
+            ],
+            applicationInsights: false,
+            isPreview: false,
+            isDeprecated: false,
+            isHidden: false,
+          },
+          {
+            displayVersion: '9.3',
+            runtimeVersion: '9.3',
+            isDefault: true,
+            minorVersions: [
+              {
+                displayVersion: '9.3.13',
+                runtimeVersion: '9.3.13.20161014',
+                isDefault: false,
+                isRemoteDebuggingEnabled: false,
+              },
+              {
+                displayVersion: '9.3.25',
+                runtimeVersion: '9.3.25.20180904',
+                isDefault: true,
+                isRemoteDebuggingEnabled: false,
+              },
+            ],
+            applicationInsights: false,
+            isPreview: false,
+            isDeprecated: false,
+            isHidden: false,
+          },
+        ],
+      },
+      {
+        name: 'java',
+        display: 'Java',
+        dependency: null,
+        majorVersions: [
+          {
+            displayVersion: 'SE',
+            runtimeVersion: 'SE',
+            isDefault: false,
+            minorVersions: [],
+            applicationInsights: false,
+            isPreview: false,
+            isDeprecated: false,
+            isHidden: false,
+          },
+        ],
+      },
+    ],
+    isDeprecated: null,
+  },
+};
