+++ server/src/stacks/webapp/2020-05-01/stacks/config/windows/node.ts
@@ -1,0 +1,392 @@
+import { WebAppConfigStack } from '../../../stack.model';
+
+export const nodeWindowsConfigStack: WebAppConfigStack = {
+  id: null,
+  name: 'node',
+  type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
+  properties: {
+    name: 'node',
+    display: 'Node',
+    dependency: null,
+    majorVersions: [
+      {
+        displayVersion: '0.6',
+        runtimeVersion: '0.6',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '0.6.20',
+            runtimeVersion: '0.6.20',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '0.8',
+        runtimeVersion: '0.8',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '0.8.2',
+            runtimeVersion: '0.8.2',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.8.19',
+            runtimeVersion: '0.8.19',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.8.26',
+            runtimeVersion: '0.8.26',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.8.27',
+            runtimeVersion: '0.8.27',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.8.28',
+            runtimeVersion: '0.8.28',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '0.10',
+        runtimeVersion: '0.10',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '0.10.5',
+            runtimeVersion: '0.10.5',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.18',
+            runtimeVersion: '0.10.18',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.21',
+            runtimeVersion: '0.10.21',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.24',
+            runtimeVersion: '0.10.24',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.28',
+            runtimeVersion: '0.10.28',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.29',
+            runtimeVersion: '0.10.29',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.31',
+            runtimeVersion: '0.10.31',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.32',
+            runtimeVersion: '0.10.32',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.40',
+            runtimeVersion: '0.10.40',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.10.5',
+            runtimeVersion: '0.10.5',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '0.12',
+        runtimeVersion: '0.12',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '0.12.0',
+            runtimeVersion: '0.12.0',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.12.2',
+            runtimeVersion: '0.12.2',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.12.3',
+            runtimeVersion: '0.12.3',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '0.12.6',
+            runtimeVersion: '0.12.6',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '4.8',
+        runtimeVersion: '4.8',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '6.5',
+        runtimeVersion: '6.5',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '6.5.0',
+            runtimeVersion: '6.5.0',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '6.9',
+        runtimeVersion: '6.9',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '6.9.1',
+            runtimeVersion: '6.9.1',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+          {
+            displayVersion: '6.9.5',
+            runtimeVersion: '6.9.5',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '6.12',
+        runtimeVersion: '6.12',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '7.10',
+        runtimeVersion: '7.10',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '7.10.1',
+            runtimeVersion: '7.10.1',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '8.1',
+        runtimeVersion: '8.1',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '8.1.4',
+            runtimeVersion: '8.1.4',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '8.4',
+        runtimeVersion: '8.4',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '8.5',
+        runtimeVersion: '8.5',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '8.9',
+        runtimeVersion: '8.9',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '8.10',
+        runtimeVersion: '8.10',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '8.11',
+        runtimeVersion: '8.11',
+        isDefault: true,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '10.0',
+        runtimeVersion: '10.0',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '10.0.0',
+            runtimeVersion: '10.0.0',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '10.6',
+        runtimeVersion: '10.6',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '10.6.0',
+            runtimeVersion: '10.6.0',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '10.14',
+        runtimeVersion: '10.14',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '10.14.0',
+            runtimeVersion: '10.14.0',
+            isDefault: false,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '10.15',
+        runtimeVersion: '10.15',
+        isDefault: false,
+        minorVersions: [
+          {
+            displayVersion: '10.15.2',
+            runtimeVersion: '10.15.2',
+            isDefault: true,
+            isRemoteDebuggingEnabled: false,
+          },
+        ],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+    ],
+    frameworks: [],
+    isDeprecated: null,
+  },
+};
