+++ server/src/stacks/webapp/2020-05-01/stacks/config/windows/php.ts
@@ -1,0 +1,66 @@
+import { WebAppConfigStack } from '../../../stack.model';
+
+export const phpWindowsConfigStack: WebAppConfigStack = {
+  id: null,
+  name: 'php',
+  type: 'Microsoft.Web/availableStacks?osTypeSelected=Windows',
+  properties: {
+    name: 'php',
+    display: 'PHP',
+    dependency: null,
+    majorVersions: [
+      {
+        displayVersion: '5.6',
+        runtimeVersion: '5.6',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '7.0',
+        runtimeVersion: '7.0',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '7.1',
+        runtimeVersion: '7.1',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '7.2',
+        runtimeVersion: '7.2',
+        isDefault: true,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+      {
+        displayVersion: '7.3',
+        runtimeVersion: '7.3',
+        isDefault: false,
+        minorVersions: [],
+        applicationInsights: false,
+        isPreview: false,
+        isDeprecated: false,
+        isHidden: false,
+      },
+    ],
+    frameworks: [],
+    isDeprecated: null,
+  },
+};
