+++ server/src/stacks/webapp/versions.ts
@@ -1,0 +1,3 @@
+export enum WebAppAPIVersions {
+  '2020-05-01' = '2020-05-01',
+}
